sidebar, avatar fixes
