+++ json-server/db.json
@@ -269,14 +269,14 @@
       "username": "admin",
       "password": "admin",
       "role": "ADMIN",
-      "avatar": "https://files.vgtimes.ru/download/posts/2017-01/thumbs/1485205515_olphimsk.jpg"
+      "avatar": "https://img.rasset.ie/001a078c-500.jpg"
     },
     {
       "id": "2",
-      "username": "Player-1",
+      "username": "Bogdan",
       "password": "Player-1",
       "role": "USER",
-      "avatar": "https://otvet.imgsmail.ru/download/181728987_1307a7339ed9ef172ee1d11c29323f2b_800.jpg"
+      "avatar": "https://upload.wikimedia.org/wikipedia/en/8/8b/HansGruber.jpeg"
     }
   ],
   "profile": [
@@ -289,7 +289,7 @@
       "country": "Russia",
       "city": "Saint-P",
       "username": "admin",
-      "avatar": "https://media-cldnry.s-nbcnews.com/image/upload/t_fit-760w,f_auto,q_auto:best/newscms/2018_50/2684041/121214-bruce-willis-die-hard-christmas-mn-1435.jpg"
+      "avatar": "https://img.rasset.ie/001a078c-500.jpg"
     },
     {
       "id": "2",
@@ -300,7 +300,7 @@
       "country": "Russia",
       "city": "Saint-P",
       "username": "Bogdan",
-      "avatar": "https://otvet.imgsmail.ru/download/181728987_1307a7339ed9ef172ee1d11c29323f2b_800.jpg"
+      "avatar": "https://upload.wikimedia.org/wikipedia/en/8/8b/HansGruber.jpeg"
     }
   ]
 }
