+++ src/entities/User/model/types/user.ts
@@ -1,3 +1,6 @@
+import { useSelector } from 'react-redux';
+import { getProfileForm } from 'entities/Profile';
+
 /**
  *    UserSchema
  *    - interface for state.
