+++ src/shared/ui/Avatar/Avatar.module.scss
@@ -13,9 +13,12 @@
 }
 
 .avatarUser {
-  width: auto;
+  width: 200%;
   height: 100%;
   margin: 0 auto;
+  background-size: cover;
+  background-position: center center;
+  background-repeat: no-repeat;
 }
 
 .clear {
