+++ src/widgets/Sidebar/models/types/sidebar.ts
@@ -1,0 +1,8 @@
+import React from 'react';
+
+export interface SidebarItemType {
+    path: string;
+    text: string;
+    Icon: React.VFC<React.SVGProps<SVGSVGElement>>;
+    authOnly?: boolean;
+}
