+++ src/widgets/Sidebar/ui/Sidebar/Sidebar.tsx
@@ -17,7 +17,8 @@ import {
 
 import BarsIcon from 'shared/assets/icons/bars-Yt.svg';
 import { SidebarItem } from 'widgets/Sidebar/ui/SidebarItem/SidebarItem';
-import { SidebarItemsList } from 'widgets/Sidebar/models/items';
+import { useSelector } from 'react-redux';
+import { getSidebarItems } from '../../models/selectors/getSidebarItems';
 import classes from './Sidebar.module.scss';
 
 interface SidebarProps {
@@ -26,6 +27,7 @@ interface SidebarProps {
 
 export const Sidebar = memo(({ className }: SidebarProps) => {
   const [collapsed, setCollapsed] = useState(true);
+  const SidebarItemsList = useSelector(getSidebarItems);
 
   const onToggle = () => {
     setCollapsed((prev) => !prev);
@@ -37,7 +39,7 @@ export const Sidebar = memo(({ className }: SidebarProps) => {
           collapsed={collapsed}
           key={item.path}
       />
-  )), [collapsed]);
+  )), [SidebarItemsList, collapsed]);
 
   return (
       <div
